+++ SipStack/SipNISTList.cpp
@@ -171,7 +171,7 @@ void CSipNISTList::DeleteAll( )
  */
 int CSipNISTList::GetSize( )
 {
-	return m_clsMap.size();
+	return (int)m_clsMap.size();
 }
 
 /**
